requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-genai
numpy
scikit-learn
streamlit